Allow failing on 2.4.0
The json-gem fails building its native extensions for Ruby 2.4.0. A
fixed version is available, but several gems depend on the older
version. These are currently:

  - activesupport 4.2.7.1
  - simplecov 0.11.1
  - rdoc 4.2.1

rdoc and simplecov can be upgraded easily. activesupport is the latest
version before 5.0, which would be most likely a bigger upgrade.

Hence we allow failing for that Ruby