Do not disrupt the user-workflow
Fixes #414

Since most DB-Errors are either setup-issues or race-conditions, I
silence them here. The setup-issues will be visible on other places and
also be logged for debugging. The race-conditions are nothing the App
can sensibly handle everytime.

So I just return "false" indicating the failed save and let the user
handle the situation according to the displayed data.