Follow the iCalendar spec more closely in the case of all-day events.
According to http://microformats.org/wiki/iCalendar and https://www.ietf.org/rfc/rfc2445.txt (section 4.8.2.2), the end date in an all-day iCalendar event is exclusive, meaning it is the earliest date after the event start that is not part of the event anymore. Since in hitobito we just consider 0:00 as a marker for all-day events, we actually need to add one day to the end date when exporting all-day events.

Refs hitobito/hitobito#606