Reduce column-width to reduce the needed index-key
The error while migrating was this:

  Mysql2::Error: Specified key was too long; max key length is 3072 bytes

Given UTF8 as encoding, 3072 bytes translate to a length of 1024 chars.
As the data being stored in the columns does currently not exceed 30 at
the moment, 100 chars should be enough for now and the forseeable
future. Since the length is validated, we will catch this one way or the
other.

If you, future reader come across this commit-message: Hi :wave:. How's
life? Keep in mind that the above limit is tied to MySQL 5.7, your
version in production might be way ahead and have a vastly superior
handling of this. If not, peace be with you and keep the sum of the
column-lengths below the maximum index prefix-length of your RDBMS
(currently 1024, but now I am repeating myself).

Oh, fun side-note: If we have already switched to utf8mb4 to support
emoji, 1 char is 4 bytes. Consider maybe not using that encoding for
data that is basically ASCII7.