Disable autosaving of an event's associated groups
Say we have a group G, its sibling group S and a person P with a role on
S but no role on G. A is allowed to create a shared event that includes
both G and S. However, `autosave: nil` or `autosave: true` would
automatically save both G and S during the creation of the event, which
will change the updated_at timestamp and set updater on both groups to
P. This can be confusing because now, on G it says that P has last
updated G, even though he shouldn't really be allowed to do so. For this
reason, we deactivate autosaving groups via the events_groups
association.