Fix displaying OAuth applications as root user
When displaying the list of OAuth applications, the sheet evaluates
which tabs to display. The path_args give nil in that case, and we end
up with a check view.can :show, nil.
Since in the special case of the root user this check passes, it is
concluded that the tabs (Info Autorisierungen Zugangstokenshould) should
be displayed, which fails with an exception.

This is fixed by explicitly checking for presence of the path_args when
determining whether to display the tabs.