MySQL has a limit on how big a row can get, but TEXT does not count
On some wagons, we have extra fields which bring the row-size close to that limit
of 65535. TEXT (and BLOB) columns are not stored directly in the row. Funnily,
this limit is indenpendent of the storage backend. InnoDB does this automatically
if a column is bigger than certain values (the concrete limit depends on the
charset). So, thank you MySQL for another gotcha.