Upgrade actionpack, keep sprockets
Sprockets 4.0 requires a manifest.js, even if you do not use it. We do
not, so I pin the version in order to avoid upgrading something we do
not use. This seems better than having a long-winded explanation in an
otherwise empty manifest.js. With our luck, someone will readd something
to the manifest.js once it exists. We chose the webpacker-route so I
want to stay true to it.

Also, resort the gems to be in alphabetical order.