Remove partially incomplete, partially undocumented redirection feature
This is to make all ways of accessing the API a little more consistent.
Also, this redirect was based on the query parameter token passing
mechanism, which I would like to deprecate in the long run, because it
has the potential to leak credentials in web server logs. The HTTP
header mechanism should therefore be preferred.
Also, I want to replace the messy token normalization (read: params
manipulation) logic in authenticatable.rb, and these are the only other
places where the manipulated params are used.

In case someone comes complaining about this, we can now tell them to
migrate their application away from user tokens, towards service tokens
or oauth tokens, and then talk about ways to add this redirect back in a
secure way.