Streetnumbers are harder to find with Sphinx
Therefore, we only use Sphinx to search for the street-name and handle
streetnumbers in ruby once we found the matching streets.

It might not be a fast, but yields more correct results and also gives
more and better feedback to the user.

Also, since we do not search number with sphinx anymore, we do not need
to index them, saving a little time and space.

  18.6 MB in 1.9 sec vs.
   5.5 MB in 0.8 sec locally.