Iron out another difference between root and non-root users
The household query would return a "must be structurally compatible"
error on the scope.or otherwise, when current_user is the root user.
The reason for this is that PersonWritables for the root user is vastly
more simple and doesn't include any JOINs or DISTINCT, while
same_address_query always includes these, regardless of who current_user
is.