Monkey-patch bug in graphiti
Graphiti's logic for assigning has_many children to an edited resource
before validation is wrong. The |= operator on arrays will check
elements for equality, and ActiveRecord objects decide their equality
by ID only. So the modified child resource is never added to the list of
children on the parent resource, and validation will therefore miss
errors in the children.

By replacing a |= b (which is the same as a = a | b) with a = b | a, we
give the modified children precedence over the existing ones, and the
nested validations on updates start working.