Allow more links to nextcloud, search them in the layer
Allow all groups to link to their nextcloud some bigger groups with
subgroups might have their own instance. On groups, show the nearest
nextcloud-instance in the layer, searching upwards the hierarchy. Do no
traverse layer-boundaries, by default. The groups that are considered
for organizing a nextcloud-instance are determined in a helper that
could be overridden in a wagon.

In general, avoid the use of FeatureGate in the views, as we have the
list of used_attributes. This is the more central way of configuring
which attributes are usable.