Deduplicate redunant information that was present multiple times
The word "Anlass" in de_CH would be "Veranstaltung" in de_DE.

While the combined word "Anlassveranstaltung" is both correct and
gender-neutral, it is also funny to read if you grew up with de_DE.
Since the "Veranstalter" is also kind of a "Verwalter", I chose the
localization "Anlassverwaltung", which is very close to the original
"Kursadministration", but a little more de_* than en_* and also not
specific to course, but all events.

Anyway, languages are a treasure-trove of fun.