Fix tests and bugs found in the process
Removes ResourceHelpers. These helpers don't simulate a realistic
situation, especially now that we don't just use in-memory permission
calculation anymore.
Avoids running the tests as root, and sets up proper roles for all
tests. Root has very special permission handling, which is not realistic
and does not provide us with real confidence that the code works in the
real world.
Renames contactable resource and ability to contact_account, to be
consistent with the terms already used in hitobito.
Adds separate abilities for reading, creating, updating and destroying
resources, and returns the correct ones (the main ability can't
currently be used to decide on any topics regarding contact accounts).
Adds missing ability for group resource.
Updates the graphiti schema, now that groups are added.