Remove authorization_spec in favor of realistic scenario specs
This spec was only testing the features of Graphiti and CanCanCan, along
with very little logic in our application_resource. It only gave us
confidence that a dummy resource (declared in the test itself) would be
authorized correctly, and this also caused problems with graphiti schema
generation. Additionally, this spec was heavily using the
resource_helpers which were previously removed, so I moved all useful
assertions to our other specs and removed this spec file.