Birthday and gender are not public
These columns have been exportable in the address export since
hitobito's inception in 2012. But by now, these columns are only
accessible to users who have the :show_details ability. Also, they are
still exportable via the full export (which is only available if one can
export the full details of all the people in the list) or the table
displays export (in which case the permission is checked for each row
separately).