Do not implement logic on class level for permissions
This has been a pain since when the Ability constant gets referenced
before any course types have been defined (e.g in the wagon.rb of a
wagon) the permissions just did not work. This happened because when the
Ability constant was loaded, the VariousAbility constant was aswell.
Thus already evaluating course types