Prevent automatic deletion of tags
We have subscription-tags which have (rightfully) a dependency
on the tags. In preventing automated deletion of tags, we
essentially prevent an error that occurs when the 'unused' tag
is actually used as a subscription-tag. This way, when removing
the tags from a person, we can add it later again without breaking
the automated subscription to a mailing-list.