Allow to delete tags
Since the automated deletion of tags after removal of the last usage
has been deactivated, we will end up with unused tags. In order to
still allow cleanup of these (and all) tags, the dependency of
subscription-tags is included into the tag-model. This allows to
delete a tag and remove it from all people (like before) and also
the subcriptions via tag of a mailing-list.