Fix selects in Event::ParticipationsController#list_entries (#2923)
The sql query generated by #list_entries is b0rken. It is only 347 lines
long when pretty printed. Somehow the timestamps selects did not work
in the jubla wagon and possibly in other wagons (but not so in core
itself).
The timestamps selects do not work when selecting "*" but do work when
selecting the columns explicitely.

Fixes hitobito/hitobito_jubla#135