Use deterministic collation when comparing with ILIKE
The benefits of our custom collation are not yielded in this case
anyway. Since our collation is nondeterministic, ILIKE (et al.) are not
supported for it. As we plan to use the case insensitive like in the
query, we can collate the content normally and use the built-in
primitives for the comparsion.