Set country to reduce flakyness
If the country is CH, it is not printed. With a random country, this
might of course happen.

Since most random countries are printed, I chose HM as the
housemate-country (because the abbreviation matches) and IO for the
single-person (because kind of reads like 1 and 0 and there is no
country SP). I realize that the IO country might vanish at since the
'Indian Overseas Territories of Britian" have been handed back to the
inhabitants. But that's a future us problem.