Remove exception which is already being handled
The MailProcessedBeforeError was a signal that needed action. In the
current mail-stack this the duplicate mail is being moved to failed and
therefore not processed again anymore. Since there is nothing to be done
by an operator or developer, this only is a signal. The fact that this
is not the responsibility of the people running, but the people owning
the the instance reduces this signal to mere noise. Unneeded noise in
a world that is too loud anyway.

Therefore, and not at all with a heavy heart, I joyfully remove this
piece of legacy that has been dutyfully ported into the present. I could
go on discussing the merits or faults of the removed code, but few will
remember and most likely no one will miss it.

Goodbye, sucker!