Silence check that points a crutch in the build-stage
We do not use real secrets in the build, it is only named so that rails
starts. In the app-image that is being run, the var is not even present
anymore. This is a very good example of a false positive. One, that we
now silence by skipping the check altogether.

That being said: please do not disclose secrets in the version-number,
mkay? That'd be great. KTHXBAI.