Fix person abilities for manageds (#3757)
We introduced an error with moving the Elternzugang to the core. The proposed fix did not address the problem properly.

Now, the permissions are again restricted in the core and relaxed in wagons if necessary (for now, only in the SAC wagon: https://github.com/hitobito/hitobito_sac_cas/pull/2101)