Clear remember_me on password reset (hitobito/hitobito_sww#310)
An error ocurred when the password of a logged in user (with
remember_me) was resetted and the user opened a page with the old
cookie.

An further improvement would be, to properly handle the bad state in the
controller.

For now, we just try to avoid the bad state and cleanup existing data:

```
Person.where(encrypted_password: nil).where.not(remember_created_at: nil).update_all remember_created_at: nil
```